Improves accuracy of measurements. Tweaks SIMD-code.

+++ test/junit5/java/ch/randelshofer/fastdoubleparser/AbstractFromByteArrayVectorizedTest.java
@@ -33,7 +33,7 @@ public abstract class AbstractFromByteArrayVectorizedTest {
         );
     }
 
-    protected abstract void testTryToParseEightDigits(String str, int expected);
+    protected abstract void testTryToParseEightDigits(String str, long expected);
 
     @TestFactory
     List<DynamicNode> dynamicTestsTryToParseEightHexDigits() {
